requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
transformers
torch
langchain
streamlit
